Add files via upload
This sketch is for Arduino Leonardo controller located in the Dragster robot. 
The sketch allows to control the battery level in Dragster robot.
The sketch measures the battery level and outputs in marquee to Troyka LED Matrix module also located in the Dragster.
Might be useful for those who plays with Dragster.
Because Dragster has the unwanted feature of constant battery power leak even when the battery is turned off. 
This results in battery becoming dead if Dragster is not recharged regularly.  
The sketch gives the visibility for the battery level and helps to avoid trouble with battery.